+++ app/Http/Controllers/DoctorAppointmentController.php
@@ -17,20 +17,32 @@ class DoctorAppointmentController extends Controller
     }
     public function show($id)
     {
-        $appointment = DoctorAppointment::findOrFail($id);
+        $appointment = DoctorAppointment::find($id);
+
+        if (!$appointment) {
+            return response()->json(['error' => 'DoctorAppointment not found'], 404);
+        }
         return $appointment;
     }
     public function showByPatient($patientId)
     {
-        return DoctorAppointment::with('patient')->where('patient_id', $patientId)->firstOrFail();
+        $appointment = DoctorAppointment::with('patient')->where('patient_id', $patientId)->first();
+        if (!$appointment) {
+            return response()->json(['error' => 'DoctorAppointment not found'], 404);
+        }
+        return $appointment;
     }
     public function showByDoctor($doctorId)
     {
-        return DoctorAppointment::with('doctor')->where('doctor_id', $doctorId)->firstOrFail();
+        $appointment = DoctorAppointment::with('doctor')->where('doctor_id', $doctorId)->first();
+        if (!$appointment) {
+            return response()->json(['error' => 'DoctorAppointment not found'], 404);
+        }
+        return $appointment;
     }
     public function store(Request $request)
     {
-        $validatedData =Validator::make( $request->all(),[
+        $validatedData = Validator::make($request->all(), [
             'patient_id' => 'required|exists:patients,patient_id',
             'doctor_id' => 'required|exists:doctors,doctor_id',
             'working_hour_id' => 'required|string',
@@ -52,25 +64,28 @@ class DoctorAppointmentController extends Controller
             'working_hour_id'  => 'required|string',
             'working_day_id'   => 'required|string',
         ]);
-    
+
         if ($validatedData->fails()) {
             return response()->json(['error' => $validatedData->errors()], 400);
         }
-    
+
         $appointment = DoctorAppointment::find($id);
-    
+
         if (!$appointment) {
             return response()->json(['error' => 'DoctorAppointment not found'], 404);
         }
-    
+
         $appointment->update($request->all());
-    
+
         return $appointment;
     }
-    
+
     public function destroy($id)
     {
-        $appointment = DoctorAppointment::findOrFail($id);
+        $appointment = DoctorAppointment::find($id);
+        if (!$appointment) {
+            return response()->json(['error' => 'DoctorAppointment not found'], 404);
+        }
         $appointment->delete();
     }
 }
